mobile: Consolidate dumpStats and terminate implementations to common/engine.cc
