+++ mobile/test/common/engine_test.cc
@@ -21,8 +21,13 @@ struct TestEngine {
   }
 
   envoy_status_t terminate() { return engine_->terminate(); }
+  bool isTerminated() const { return engine_->isTerminated(); }
 
-  ~TestEngine() { engine_->terminate(); }
+  ~TestEngine() {
+    if (!engine_->isTerminated()) {
+      engine_->terminate();
+    }
+  }
 };
 
 class EngineTest : public testing::Test {
@@ -54,6 +59,7 @@ TEST_F(EngineTest, EarlyExit) {
   ASSERT_TRUE(test_context.on_engine_running.WaitForNotificationWithTimeout(absl::Seconds(10)));
 
   ASSERT_EQ(engine_->terminate(), ENVOY_SUCCESS);
+  ASSERT_TRUE(engine_->isTerminated());
   ASSERT_TRUE(test_context.on_exit.WaitForNotificationWithTimeout(absl::Seconds(10)));
 
   engine_->engine_->startStream(0, {}, false);
@@ -77,15 +83,14 @@ TEST_F(EngineTest, AccessEngineAfterInitialization) {
   ASSERT_TRUE(test_context.on_engine_running.WaitForNotificationWithTimeout(absl::Seconds(10)));
 
   absl::Notification getClusterManagerInvoked;
-  envoy_data stats_data;
   // Running engine functions should work because the engine is running
-  EXPECT_EQ(ENVOY_SUCCESS, engine_->engine_->dumpStats(&stats_data));
-  release_envoy_data(stats_data);
+  EXPECT_NE("", engine_->engine_->dumpStats());
 
   engine_->terminate();
+  ASSERT_TRUE(engine_->isTerminated());
 
   // Now that the engine has been shut down, we no longer expect scheduling to work.
-  EXPECT_EQ(ENVOY_FAILURE, engine_->engine_->dumpStats(&stats_data));
+  EXPECT_EQ("", engine_->engine_->dumpStats());
 
   engine_.reset();
 }
